Servicio de Guardar articulos

+++ src/main/java/com/inventario/dao/ICategoriaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.inventario.dto.CategiraDto;
 
-public interface ICategoriaDao extends CrudRepository<CategiraDto, Long> {// es una clase generica
+public interface ICategoriaDao extends CrudRepository<CategiraDto, Integer> {// es una clase generica
 
 }
